Add LootTableLoadEvent, methods on LootPool/LootTable
The mixin to LootManager is messy, and I'm open to alternative ways to
doing it, even if it involves scrapping the 'forge way' of doing it and
doing it a different way. For now, I'm going to get this up so that
bikeshedding can happen.

I'm using the "net.patchworkmc.api.*.Forge<classname>" pattern for
interfaces that describe public methods added by Forge for use by mods
in this commit. This is up for bikeshedding.

This is untested other than starting up Minecraft and opening a world.