Make slightly less hacky!
The idea is to add an inject in the same place as the method we would be
redirecting (but are instead cancelling) so that we can capture locals,
and grab the builder that we would otherwise be missing.

The downside is that we are effectively overwriting the lambda that
vanilla uses in favor of our own.