Add LootTableLoadEvent, methods on LootPool/LootTable (#113)
* Add LootTableLoadEvent, methods on LootPool/LootTable

The mixin to LootManager is messy, and I'm open to alternative ways to
doing it, even if it involves scrapping the 'forge way' of doing it and
doing it a different way. For now, I'm going to get this up so that
bikeshedding can happen.

I'm using the "net.patchworkmc.api.*.Forge<classname>" pattern for
interfaces that describe public methods added by Forge for use by mods
in this commit. This is up for bikeshedding.

This is untested other than starting up Minecraft and opening a world.

* Fix LocalCapture.PRINT that I forgot about

* Implement FMLServerStoppedEvent

* Dispatch from the god classes for the events added in this PR

* prevent memory leak

* fixup! checkstype

* Make slightly less hacky!

The idea is to add an inject in the same place as the method we would be
redirecting (but are instead cancelling) so that we can capture locals,
and grab the builder that we would otherwise be missing.

The downside is that we are effectively overwriting the lambda that
vanilla uses in favor of our own.

* Make the implicit "overwrite" explicit.

* Make requested changes.

Mainly adding Override annotations to methods for interfaces, but a few
other things as well.