99% of the way to the alternative implementation I've been talking about
Only issue is: once again, the lambda is static, and that *doesn't work
for this* because we need the lootManager to fire the event.

I do have a couple ideas revolving around doing something hacky with
JsonElement, and doing something hacky with ImmutableMap.Builder, but
this is my current status.